add utils to configurator
