+++ openprocurement/auctions/core/adapters.py
@@ -4,3 +4,4 @@ from openprocurement.api.adapters import ContentConfigurator
 class AuctionConfigurator(ContentConfigurator):
     name = 'Auction Configurator'
     model = None
+    award_model = None
