+++ openprocurement/auctions/core/plugins/awarding_1_0/views/award.py
@@ -12,7 +12,6 @@ from openprocurement.api.utils import (
 from openprocurement.auctions.core.utils import (
     apply_patch,
     save_auction,
-    add_next_award,
 )
 from openprocurement.auctions.core.validation import (
     validate_award_data,
@@ -309,7 +308,7 @@ class AuctionAwardResource(APIResource):
                 'date': get_now(),
                 'items': [i for i in auction.items if i.relatedLot == award.lotID ],
                 'contractID': '{}-{}{}'.format(auction.auctionID, self.server_id, len(auction.contracts) + 1) }))
-            add_next_award(self.request)
+            self.request.content_configurator.add_award()
         elif award_status == 'active' and award.status == 'cancelled':
             now = get_now()
             if award.complaintPeriod.endDate > now:
@@ -322,10 +321,10 @@ class AuctionAwardResource(APIResource):
             for i in auction.contracts:
                 if i.awardID == award.id:
                     i.status = 'cancelled'
-            add_next_award(self.request)
+            self.request.content_configurator.add_award()
         elif award_status == 'pending' and award.status == 'unsuccessful':
             award.complaintPeriod.endDate = calculate_business_date(get_now(), STAND_STILL_TIME, auction, True)
-            add_next_award(self.request)
+            self.request.content_configurator.add_award()
         elif award_status == 'unsuccessful' and award.status == 'cancelled' and any([i.status in ['claim', 'answered', 'pending', 'resolved'] for i in award.complaints]):
             if auction.status == 'active.awarded':
                 auction.status = 'active.qualification'
@@ -347,7 +346,7 @@ class AuctionAwardResource(APIResource):
             for i in auction.contracts:
                 if i.awardID in cancelled_awards:
                     i.status = 'cancelled'
-            add_next_award(self.request)
+            self.request.content_configurator.add_award()
         elif self.request.authenticated_role != 'Administrator' and not(award_status == 'pending' and award.status == 'pending'):
             self.request.errors.add('body', 'data', 'Can\'t update award in current ({}) status'.format(award_status))
             self.request.errors.status = 403
