+++ openprocurement/auctions/core/plugins/awarding_2_0/utils.py
@@ -1,5 +1,59 @@
 from openprocurement.api.utils import get_now
 from openprocurement.api.models import TZ
+from barbecue import chef
+
+
+def create_awards_dgf(request):
+    auction = request.validated['auction']
+    auction.status = 'active.qualification'
+    now = get_now()
+    auction.awardPeriod = type(auction).awardPeriod({'startDate': now})
+
+    bids = chef(auction.bids, auction.features or [], [], True)
+
+    for i, status in enumerate(['pending.verification', 'pending.waiting']):
+        bid = bids[i].serialize()
+        award = type(auction).awards.model_class({
+            '__parent__': request.context,
+            'bid_id': bid['id'],
+            'status': status,
+            'date': now,
+            'value': bid['value'],
+            'suppliers': bid['tenderers'],
+            'complaintPeriod': {'startDate': now}
+        })
+        if bid['status'] == 'invalid':
+            award.status = 'unsuccessful'
+            award.complaintPeriod.endDate = now
+        if award.status == 'pending.verification':
+            award.signingPeriod = award.paymentPeriod = award.verificationPeriod = {'startDate': now}
+            request.response.headers['Location'] = request.route_url('{}:Auction Awards'.format(auction.procurementMethodType), auction_id=auction.id, award_id=award['id'])
+        auction.awards.append(award)
+
+
+def create_awards_insider(request):
+    auction = request.validated['auction']
+    auction.status = 'active.qualification'
+    now = get_now()
+    auction.awardPeriod = type(auction).awardPeriod({'startDate': now})
+    valid_bids = [bid for bid in auction.bids if bid['status'] != 'invalid']
+    bids = chef(valid_bids, auction.features or [], [], True)
+
+    for bid, status in zip(bids, ['pending.verification', 'pending.waiting']):
+        bid = bid.serialize()
+        award = type(auction).awards.model_class({
+            '__parent__': request.context,
+            'bid_id': bid['id'],
+            'status': status,
+            'date': now,
+            'value': bid['value'],
+            'suppliers': bid['tenderers'],
+            'complaintPeriod': {'startDate': now}
+        })
+        if award.status == 'pending.verification':
+            award.signingPeriod = award.paymentPeriod = award.verificationPeriod = {'startDate': now}
+            request.response.headers['Location'] = request.route_url('{}:Auction Awards'.format(auction.procurementMethodType), auction_id=auction.id, award_id=award['id'])
+        auction.awards.append(award)
 
 
 def switch_to_next_award(request):
